requirements.txt
Flask==0.12.2
google-cloud-firestore==0.29.0